introduce pipeline helpers
