+++ mlflow/pipelines/utils/metrics.py
@@ -1,107 +1,15 @@
 import logging
 import importlib
 import sys
-from typing import List, Dict, Optional
+from typing import List, Dict
 
 from mlflow.exceptions import MlflowException, BAD_REQUEST
-from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
+from mlflow.pipelines.helpers import PipelineHelper, PipelineMetric
 
 _logger = logging.getLogger(__name__)
 
 
-class PipelineMetric:
-
-    _KEY_METRIC_NAME = "name"
-    _KEY_METRIC_GREATER_IS_BETTER = "greater_is_better"
-    _KEY_CUSTOM_FUNCTION = "function"
-
-    def __init__(self, name: str, greater_is_better: bool, custom_function: Optional[str] = None):
-        self.name = name
-        self.greater_is_better = greater_is_better
-        self.custom_function = custom_function
-
-    @classmethod
-    def from_custom_metric_dict(cls, custom_metric_dict):
-        metric_name = custom_metric_dict.get(PipelineMetric._KEY_METRIC_NAME)
-        greater_is_better = custom_metric_dict.get(PipelineMetric._KEY_METRIC_GREATER_IS_BETTER)
-        custom_function = custom_metric_dict.get(PipelineMetric._KEY_CUSTOM_FUNCTION)
-        if (metric_name, greater_is_better, custom_function).count(None) > 0:
-            raise MlflowException(
-                f"Invalid custom metric definition: {custom_metric_dict}",
-                error_code=INVALID_PARAMETER_VALUE,
-            )
-
-        return cls(
-            name=metric_name, greater_is_better=greater_is_better, custom_function=custom_function
-        )
-
-
-BUILTIN_CLASSIFICATION_PIPELINE_METRICS = [
-    PipelineMetric(name="true_negatives", greater_is_better=True),
-    PipelineMetric(name="false_positives", greater_is_better=False),
-    PipelineMetric(name="false_negatives", greater_is_better=False),
-    PipelineMetric(name="true_positives", greater_is_better=True),
-    PipelineMetric(name="recall", greater_is_better=True),
-    PipelineMetric(name="precision", greater_is_better=True),
-    PipelineMetric(name="f1_score", greater_is_better=True),
-    PipelineMetric(name="accuracy_score", greater_is_better=True),
-    PipelineMetric(name="log_loss", greater_is_better=False),
-    PipelineMetric(name="roc_auc", greater_is_better=True),
-    PipelineMetric(name="precision_recall_auc", greater_is_better=True),
-]
-
-BUILTIN_REGRESSION_PIPELINE_METRICS = [
-    PipelineMetric(name="mean_absolute_error", greater_is_better=False),
-    PipelineMetric(name="mean_squared_error", greater_is_better=False),
-    PipelineMetric(name="root_mean_squared_error", greater_is_better=False),
-    PipelineMetric(name="max_error", greater_is_better=False),
-    PipelineMetric(name="mean_absolute_percentage_error", greater_is_better=False),
-]
-
-
-def _get_error_fn(tmpl: str):
-    """
-    :param tmpl: The template kind, e.g. `regression/v1`.
-    :return: The error function for the provided template.
-    """
-    if tmpl == "regression/v1":
-        return lambda predictions, targets: predictions - targets
-    raise MlflowException(
-        f"No error function for template kind {tmpl}",
-        error_code=INVALID_PARAMETER_VALUE,
-    )
-
-
-def _get_model_type_from_template(tmpl: str) -> str:
-    """
-    :param tmpl: The template kind, e.g. `regression/v1`.
-    :return: A model type literal compatible with the mlflow evaluation service, e.g. regressor.
-    """
-    if tmpl == "regression/v1":
-        return "regressor"
-    raise MlflowException(
-        f"No model type for template kind {tmpl}",
-        error_code=INVALID_PARAMETER_VALUE,
-    )
-
-
-def _get_builtin_metrics(tmpl: str) -> str:
-    """
-    :param tmpl: The template kind, e.g. `regression/v1`.
-    :return: The builtin metrics for the mlflow evaluation service for the model type for
-    this template.
-    """
-    if tmpl == "regression/v1":
-        return BUILTIN_REGRESSION_PIPELINE_METRICS
-    elif tmpl == "classification/v1":
-        return BUILTIN_CLASSIFICATION_PIPELINE_METRICS
-    raise MlflowException(
-        f"No builtin metrics for template kind {tmpl}",
-        error_code=INVALID_PARAMETER_VALUE,
-    )
-
-
-def _get_custom_metrics(step_config: Dict) -> List[Dict]:
+def _get_custom_metrics(step_config: Dict, pipeline_helper: PipelineHelper) -> List[Dict]:
     """
     :param: Configuration dictionary for the train or evaluate step.
     :return: A list of custom metrics defined in the specified configuration dictionary,
@@ -112,9 +20,7 @@ def _get_custom_metrics(step_config: Dict) -> List[Dict]:
         PipelineMetric.from_custom_metric_dict(metric_dict) for metric_dict in custom_metric_dicts
     ]
     custom_metric_names = {metric.name for metric in custom_metrics}
-    builtin_metric_names = {
-        metric.name for metric in _get_builtin_metrics(step_config.get("template_name"))
-    }
+    builtin_metric_names = {metric.name for metric in pipeline_helper.builtin_metrics()}
     overridden_builtin_metrics = custom_metric_names.intersection(builtin_metric_names)
     if overridden_builtin_metrics:
         _logger.warning(
